refactor: Update UsersTable component to improve user feedback and layout
